Refactor thin client. Current structure is this: |- README.md |- requirements.txt |- .gitignore |- .travis.yml |- thin_client |   |-- __init__.py |   |-- thin_client_main.py |   |-- networking.py |   |-- vlc_addon.py |   |-- settings.py |   |-- test |       |-- test_thin_client.py |       |-- test_networking.py
